cull particles connectedDreamBlock
its definitely not the proper way to do it. i didn't use Celeste.Mod.CullHelper.IsRectangleVisible, and i tried to get the texture to cull it based on that but it still lagged a bit, so i left it this way. it looks easy so please someone do it please :)